Test1_AgentsAndRAG
This code creates 2 agents:
- Agent 1, creates a json file. It takes as an input natural language of the rule creation and it gives as an output a json file separating the conditions and the actions of the request. This agent helps understanding what are the conditions in order to execute the following action, this will help a future agent to validate the rules in case the conditions are met and the actions do not contradict other rules. 
Agent 1 has a RAG implemented, at the moment it just works with 1 document and this document is saved in a folder in Drive (make sure to change the file path to your local one). 
In this document some basic business considerations have been described, but this has to be improved, as well as the preprocessing of the document. 
The output of Agent 1 is passed as an input of Agent 2.

- Agent 2 takes the json file generated from Agent 1 and uses Gemini knowledge to generate a drl rule with the correct drools syntax. Obviously this will need to be fine tuned in order to use the correct fields and entities for our rules but this has to be defined yet.
Agent 2 outputs some text that is later converted into a .drl file and stored locally (path can be specified).
This agent is not connected to any RAG system at the moment.